+++ CHANGELOG.md
@@ -2,12 +2,13 @@
 # Change Log
 
 ## To Do
-- Integrate an OAuth2 library.
-- Wrap the API end-points with client methods.
+- Wrap the remaining API end-points with client methods.
 - Add specs.
 - Add docs.
 - Automatics test suite run. (Codeship/Travis)
 
 ## [Unreleased](unreleased)
 ### Added
-- No recent changes.
+- Integrate an OAuth2 library.
+- Wrappers for exchange rates end-points.
+- Flexible manager for Oauth2 and API calls.